requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==6.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
